third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-multiline-title-expected.txt

PASS picker.selectElement_.options[0].title is "bar\nbaz"
PASS successfullyParsed is true

TEST COMPLETE

foo
